pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel", "numpy"]
build-backend = "setuptools.build_meta"

[project]
version = "0.0.0"
name = "QTradeX"
description = "AI-powered SDK featuring algorithmic trading, backtesting, deployment on 100+ exchanges, and multiple optimization engines."
readme = {file = "README.md", content-type = "text/markdown"}
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "squidKid-deluxe"}]  # Optional; adjust as needed
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
]
dependencies = [
    "ccxt",
    "jsonpickle",
    "cachetools",
    "yfinance",
    "finance-datareader",
    "numpy",
    "matplotlib",
    "scipy",
    "ttkbootstrap",
    "pandas_ta",
    "websocket-client",
    "coincurve",
    "python-dotenv",
]





[project.urls]
Homepage = "https://github.com/squidKid-deluxe/QTradeX-Algo-Trading-SDK"
Issues = "https://github.com/squidKid-deluxe/QTradeX-Algo-Trading-SDK/issues"

[project.scripts]
qtradex-tune-manager = "qtradex.core.tune_manager:main"

[tool.setuptools]
packages = {find = {where = ["."]}}
include-package-data = true

[tool.setuptools.package-data]
qtradex = ["*.pxd"]
